+++ src/clorag/utils/__init__.py
@@ -1,5 +1,12 @@
 """Utility functions and helpers."""
 
 from clorag.utils.anonymizer import AnonymizationContext, TextAnonymizer
+from clorag.utils.tokenizer import count_tokens, get_encoder, truncate_to_tokens
 
-__all__ = ["AnonymizationContext", "TextAnonymizer"]
+__all__ = [
+    "AnonymizationContext",
+    "TextAnonymizer",
+    "count_tokens",
+    "get_encoder",
+    "truncate_to_tokens",
+]